requirements.txt
discord.py >= 1.7.3
toml >= 0.10.2
Pillow >= 9.2.0
requests >= 2.27.1